// packages/component/src/@types/charting_library.d.ts
export declare type Nominal<T, Name extends string> = T & {
  /* eslint-disable-next-line jsdoc/require-jsdoc */ [Symbol.species]: Name;
};
// ...
export interface Bar {
  /** Bar time.
   * Amount of **milliseconds** since Unix epoch start in **UTC** timezone.
   * `time` for daily bars is expected to be a trading day (not session start day) at 00:00 UTC.
   * Charting Library adjusts time according to `session` from {@link LibrarySymbolInfo}.
   */
  time: number;
  /** Opening price */
  open: number;
  /** High price */
  high: number;
  /** Low price */
  low: number;
  /** Closing price */
  close: number;
  /** Trading Volume */
  volume?: number;
}
// ...
export interface CurrencyItem {
  /** Unique ID */
  id: string;
  /** Currency code. @example `USD` */
  code: string;
  /** URL to an image of the currency. SVG logos are preferable, but please make sure that the provided image fits nicely in the currency select control (for raster images the expected size is 24x24px). */
  logoUrl?: string;
  /** Description for the currency */
  description?: string;
}
// ...
export interface DatafeedConfiguration {
  /**
   * List of exchange descriptors.
   * An empty array leads to an absence of the exchanges filter in the Symbol Search list.
   * Use `value=''` if you wish to include all exchanges.
   */
  exchanges?: Exchange[];
  /**
   * List of supported resolutions. Resolution string format is described here: {@link ResolutionString}
   * Setting this property to `undefined` or an empty array will result in the resolution widget
   * displaying the content.
   *
   * @example
   * `["1", "15", "240", "D", "6M"]` will give you "1 minute, 15 minutes, 4 hours, 1 day, 6 months" in resolution widget.
   */
  supported_resolutions?: ResolutionString[];
  /**
   * Supported unit groups. Each group can have several unit objects.
   *
   * @example
   * ```javascript
   * {
   *     weight: [
   *         { id: 'kg', name: 'kg', description: 'Kilograms' },
   *         { id: 'lb', name: 'lb', description: 'Pounds'},
   *     ]
   * }
   * ```
   */
  units?: Record<string, Unit[]>;
  /**
   * Supported currencies for currency conversion.
   *
   * When a currency code is supplied as a string then the library will automatically convert it to `{ id: value, code: value }` object.
   *
   * @example `["USD", "EUR", "GBP"]`
   * @example `[{ id: "USD", code: "USD", description: "$" }, { id: "EUR", code: "EUR", description: "€" }]`
   */
  currency_codes?: (string | CurrencyItem)[];
  /** Does the datafeed supports marks on bars (`true`), or not (`false | undefined`). */
  supports_marks?: boolean;
  /** Set this one to `true` if your datafeed provides server time (unix time). It is used to adjust Countdown on the Price scale. */
  supports_time?: boolean;
  /** Does the datafeed supports marks on the timescale (`true`), or not (`false | undefined`). */
  supports_timescale_marks?: boolean;
  /**
   * List of filter descriptors.
   *
   * Setting this property to `undefined` or an empty array leads to the absence of filter types in Symbol Search list. Use `value = ''` if you wish to include all filter types.
   * `value` within the descriptor will be passed as `symbolType` argument to {@link IDatafeedChartApi.searchSymbols}
   */
  symbols_types?: DatafeedSymbolType[];
  /**
   * Set it if you want to group symbols in the symbol search.
   * Represents an object where keys are symbol types {@link SymbolType} and values are regular expressions (each regular expression should divide an instrument name into 2 parts: a root and an expiration).
   *
   * Sample:
   * ```
   * {
   *   "futures": `/^(.+)([12]!|[FGHJKMNQUVXZ]\d{1,2})$/`,
   *   "stock": `/^(.+)([12]!|[FGHJKMNQUVXZ]\d{1,2})$/`,
   * }
   * ```
   * It will be applied to the instruments with futures and stock as a type.
   */
  symbols_grouping?: Record<string, string>;
}
// ...
export interface DatafeedSymbolType {
  /** Name of the symbol type */
  name: string;
  /** Value to be passed as the `symbolType` argument to `searchSymbols` */
  value: string;
}
// ...
export interface Exchange {
  /** Value to be passed as the `exchange` argument to `searchSymbols` */
  value: string;
  /** Name of the exchange */
  name: string;
  /** Description of the exchange */
  desc: string;
}
// ...
export interface IExternalDatafeed {
  /**
   * This call is intended to provide the object filled with the configuration data.
   * Charting Library assumes that you will call the callback function and pass your datafeed {@link DatafeedConfiguration} as an argument.
   *
   * @param  {OnReadyCallback} callback - callback to return your datafeed configuration ({@link DatafeedConfiguration}) to the library.
   */
  onReady(callback: OnReadyCallback): void;
}
// ...
export interface LibrarySymbolInfo {
// ...
  description: string;
// ...
  timezone: Timezone;
// ...
  pricescale: number;
// ...
}
// ...
export interface QuoteErrorData extends QuoteDataResponse {
  /** @inheritDoc */
  s: "error";
  /** @inheritDoc */
  v: object;
}
/** Quote Data Ok Response */
export interface QuoteOkData extends QuoteDataResponse {
  /** @inheritDoc */
  s: "ok";
  /** @inheritDoc */
  v: DatafeedQuoteValues;
}
// ...
export interface Unit {
  /** Unique ID */
  id: string;
  /** Short name */
  name: string;
  /** Description */
  description: string;
}
// ...
export type CustomTimezones =
  | "Africa/Cairo"
  | "Africa/Johannesburg"
  | "Africa/Lagos"
  | "Africa/Nairobi"
  | "Africa/Tunis"
  | "America/Argentina/Buenos_Aires"
  | "America/Bogota"
  | "America/Caracas"
  | "America/Chicago"
  | "America/El_Salvador"
  | "America/Juneau"
  | "America/Lima"
  | "America/Los_Angeles"
  | "America/Mexico_City"
  | "America/New_York"
  | "America/Phoenix"
  | "America/Santiago"
  | "America/Sao_Paulo"
  | "America/Toronto"
  | "America/Vancouver"
  | "Asia/Almaty"
  | "Asia/Ashkhabad"
  | "Asia/Bahrain"
  | "Asia/Bangkok"
  | "Asia/Chongqing"
  | "Asia/Colombo"
  | "Asia/Dubai"
  | "Asia/Ho_Chi_Minh"
  | "Asia/Hong_Kong"
  | "Asia/Jakarta"
  | "Asia/Jerusalem"
  | "Asia/Karachi"
  | "Asia/Kathmandu"
  | "Asia/Kolkata"
  | "Asia/Kuwait"
  | "Asia/Manila"
  | "Asia/Muscat"
  | "Asia/Nicosia"
  | "Asia/Qatar"
  | "Asia/Riyadh"
  | "Asia/Seoul"
  | "Asia/Shanghai"
  | "Asia/Singapore"
  | "Asia/Taipei"
  | "Asia/Tehran"
  | "Asia/Tokyo"
  | "Asia/Yangon"
  | "Atlantic/Reykjavik"
  | "Australia/Adelaide"
  | "Australia/Brisbane"
  | "Australia/Perth"
  | "Australia/Sydney"
  | "Europe/Amsterdam"
  | "Europe/Athens"
  | "Europe/Belgrade"
  | "Europe/Berlin"
  | "Europe/Bratislava"
  | "Europe/Brussels"
  | "Europe/Bucharest"
  | "Europe/Budapest"
  | "Europe/Copenhagen"
  | "Europe/Dublin"
  | "Europe/Helsinki"
  | "Europe/Istanbul"
  | "Europe/Lisbon"
  | "Europe/London"
  | "Europe/Luxembourg"
  | "Europe/Madrid"
  | "Europe/Malta"
  | "Europe/Moscow"
  | "Europe/Oslo"
  | "Europe/Paris"
  | "Europe/Riga"
  | "Europe/Rome"
  | "Europe/Stockholm"
  | "Europe/Tallinn"
  | "Europe/Vilnius"
  | "Europe/Warsaw"
  | "Europe/Zurich"
  | "Pacific/Auckland"
  | "Pacific/Chatham"
  | "Pacific/Fakaofo"
  | "Pacific/Honolulu"
  | "Pacific/Norfolk"
  | "US/Mountain";
// ...
export type IBasicDataFeed = IDatafeedChartApi & IExternalDatafeed;
// ...
export type OnReadyCallback = (configuration: DatafeedConfiguration) => void;
// ...
export type QuoteData = QuoteOkData | QuoteErrorData;
/**
 * Callback to provide Quote data.
 * @param  {QuoteData[]} data - Quote Data
 */
export type QuotesCallback = (data: QuoteData[]) => void;
/**
 * Error callback for quote data request.
 * @param  {QuoteData[]} reason - message describing the reason for the error
 */
export type QuotesErrorCallback = (reason: string) => void;
// ...
export type ResolutionString = Nominal<string, "ResolutionString">;
// ...
export type SubscribeBarsCallback = (bar: Bar) => void;
// ...
export type Timezone = "Etc/UTC" | CustomTimezones;